chore: move the scheduler internal components out of internal dir
